put files name back how they were
