+++ apps/newsletters-api/src/services/storage/index.ts
@@ -14,7 +14,10 @@ import {
 } from '@newsletters-nx/newsletters-data-client';
 import newslettersData from '../../../static/newsletters.local.json';
 import { isUsingInMemoryStorage } from '../../apiDeploymentSettings';
-import { makeEmailEnvInfo, makeSesClient } from '../email-client/make-client';
+import {
+	makeEmailEnvInfo,
+	makeSesClient,
+} from '../notifications/email-service';
 import { makeInMemoryStorageInstance } from './inMemoryStorageInstance';
 import {
 	getS3NewsletterStore,
